done1

+++ app/Http/Controllers/ArticleController.php
@@ -1,14 +1,15 @@
-<?php namespace App\Http\Controllers;
+<?php
+
+namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
-
 use App\Article;
 use Carbon\Carbon;
 use App\Http\Requests\ArticleRequest;
 use Auth;
+use Gate;
 
 class ArticleController extends Controller
 {
@@ -81,6 +82,8 @@ class ArticleController extends Controller
         //$article = Article::findOrFail($id);//if(is_null($user)) abort(404);
         //dd($article->published_at);
         //dd($article->created_at->addDays(8)->diffForHumans());//year//month//addDays(8)->format('Y-m')//diffForHumans()
+        //auth()->logout();//or change the users id
+        //auth()->loginUsingId(11);//tmp
         return view('articles.article', compact('article'));
     }
 
@@ -92,6 +95,19 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)//$id
     {
+         // if(Gate::denies('edit', $article))//allows //edit-article if use boot
+         // {
+         //     abort(403, 'Sorry, not sorry.');
+         // }
+        //$this->authorize('edit-article', $article);
+        // if(auth()->user()->can('edit-article', $article))//cannot
+        // {
+        //     return 'You can see this.';
+        // }
+        if(auth()->user()->cannot('edit', $article))
+        {
+            return redirect('article');
+        }
         //$article = Article::findOrFail($id);
         $tags = \App\Tag::lists('name', 'id');
         return view('articles.edit', compact('article', 'tags'));
@@ -106,6 +122,11 @@ class ArticleController extends Controller
      */
     public function update(ArticleRequest $request, Article $article)//$id
     {
+        //if($this->authorize('edit-article', $article))
+        if($request->user()->cannot('edit', $article))
+        {
+            return redirect('article');
+        }
         //$article = Article::findOrFail($id);
         $article->update($request->all());
         //$article->tags()->sync($request->input('tag_list'));
@@ -120,6 +141,10 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        if(auth()->user()->cannot('edit', $article))
+        {
+            return redirect('article');
+        }
         $article->tags()->detach();
         $article->delete();
         flash('Your article has been deleted!');
